rift: Update blobs from long analysis observations.
Add a unique blob id that stays the same as blobs are tracked
using motion flow, and use them to match blobs at the end
of a long analysis that found good pose(s). Transfer the
LED labels onto the most recent frame data for the next
fast analysis cycle to use.